Plug the C# UDP server into the project

+++ Assets/Scripts/Network/NetworkManager.cs
@@ -1,8 +1,12 @@
 ﻿using GameProtocol;
+using spacial_dogfight_protocol.Commands;
+using spacial_dogfight_protocol.Messages;
 using System;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.Net;
 using System.Net.Sockets;
+using System.Text;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
@@ -32,44 +36,40 @@ public class NetworkManager : MonoBehaviour {
     public ClientStatus Status { get; private set; }
     public string IpAddress { get; set; }
 
+    public Dictionary<int, List<NetworkCallback>> updateCallbacks;
+
     private UdpClient _client;
     private UdpClient _broadcastClient;
-
-
     protected Process _serverProcess;
 
-	void Start () {
-		
-	}
-	
-	void Update () {
-		
-	}
+    private bool commandSend = false;
 
     private void OnDestroy() {
-        _serverProcess.Kill();
+        if(_serverProcess != null)
+            _serverProcess.Kill();
     }
 
     public void StartAsServer() {
         _serverProcess = new Process();
-        _serverProcess.StartInfo.FileName = "E:\\repositories\\spacial_dogfight\\Assets\\Externals\\Server\\spacial_dogfight_server.exe";
+        _serverProcess.StartInfo.FileName = Application.dataPath + "\\Externals\\Server\\spacial_dogfight_server.exe";
         if (!_serverProcess.Start())
             UnityEngine.Debug.Log("Cannot start server");
         else
             UnityEngine.Debug.Log("Server launched");
 
-        Initialize(IPAddress.Loopback.ToString(), 1100);
+        Initialize("192.168.0.19", this.ServerPort);
         SceneManager.LoadScene("Main");
-        GetComponent<CameraManager>().enabled = true;
+        //GetComponent<CameraManager>().enabled = true;
         Connect();
     }
 
     public void StartAsClient() {
-        Initialize(IpAddress, 1100);
+        Initialize(ServerAddress, this.ServerPort);
         Connect();
     }
 
     public NetworkManager() {
+        updateCallbacks = new Dictionary<int, List<NetworkCallback>>();
         Status = ClientStatus.NotInitialized;
         _instance = this;
     }
@@ -90,30 +90,40 @@ public class NetworkManager : MonoBehaviour {
         }
     }
 
+    public void AddUpdateCallback(int category, int code, NetworkCallback callback) {
+        var opcode = category << 8 | code;
+        if (!updateCallbacks.ContainsKey(opcode))
+            updateCallbacks[opcode] = new List<NetworkCallback>();
+
+        updateCallbacks[opcode].Add(callback);
+    }
+
     public void Connect() {
         Status = ClientStatus.Connecting;
 
         _client = new UdpClient();
         _client.EnableBroadcast = true;
         _client.Connect(IPAddress.Parse(ServerAddress), Port);
-        _broadcastClient = new UdpClient();
-        _broadcastClient.EnableBroadcast = true;
-        _broadcastClient.Connect(new IPEndPoint(IPAddress.Parse("192.168.0.255"), Port));
+        //_broadcastClient = new UdpClient();
+        //_broadcastClient.EnableBroadcast = true;
+        //_broadcastClient.Connect(new IPEndPoint(IPAddress.Parse("192.168.0.255"), Port));
         Status = ClientStatus.Connected;
 
         if (Status != ClientStatus.Disconnecting) {
+            UnityEngine.Debug.Log("Connected");
             _client.BeginReceive(new AsyncCallback(ReceiveCallback), null);
-            _broadcastClient.BeginReceive(new AsyncCallback(BroadCastCallback), null);
+            //_broadcastClient.BeginReceive(new AsyncCallback(BroadCastCallback), null);
         }
         else {
-            _broadcastClient.Close();
+            //_broadcastClient.Close();
             _client.Close();
             Status = ClientStatus.Disconnected;
         }
     }
 
     public void ReceiveCallback(IAsyncResult result) {
-        IPEndPoint localEndPoint = new IPEndPoint(IPAddress.Any, Port);
+        UnityEngine.Debug.Log("Receive");
+        IPEndPoint localEndPoint = new IPEndPoint(IPAddress.Parse("192.168.0.19"), Port);
         var receivedBuffer = _client.EndReceive(result, ref localEndPoint);
         Update update = GameProtocol.Update.GetUpdate(ref receivedBuffer);
         if (update == null) {
@@ -128,7 +138,8 @@ public class NetworkManager : MonoBehaviour {
             return;
         }
 
-        update.ProcessUpdate();
+        var opcode = update.Category << 8 | update.Code;
+        updateCallbacks[opcode].ForEach(x => x.ProcessUpdate(update));
 
         if (Status != ClientStatus.Disconnecting) {
             _client.BeginReceive(new AsyncCallback(ReceiveCallback), null);
@@ -157,7 +168,8 @@ public class NetworkManager : MonoBehaviour {
             return;
         }
 
-        update.ProcessUpdate();
+        var opcode = update.Category << 8 | update.Code;
+        updateCallbacks[opcode].ForEach(x => x.ProcessUpdate(update));
 
         if (Status != ClientStatus.Disconnecting) {
             _broadcastClient.BeginReceive(new AsyncCallback(ReceiveCallback), null);
